Split k8s tasks out from azure one
