+++ tasks/__init__.py
@@ -1,6 +1,7 @@
 from invoke import Collection
 
 from . import cluster
+from . import k8s
 from . import monitor
 from . import storage
 from . import uk8s
@@ -8,6 +9,7 @@ from . import vm
 
 ns = Collection(
     cluster,
+    k8s,
     monitor,
     storage,
     uk8s,
